+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 * Add disabled by default `RSpec/AlignRightLetBrace`. ([@backus][])
 * Add `RSpec/LetBeforeExamples` cop. ([@Darhazer][])
 * Add `RSpec/MultipleSubjects` cop. ([@backus][])
+* Add `RSpec/ReturnFromStub` cop. ([@Darhazer][])
 
 ## 1.15.1 (2017-04-30)